backend/pyproject.toml
[project]
name = "quantclass-backend"
version = "0.2.3"
description = "QuantClass Smart Browser Extension - Backend Service"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "aiosqlite>=0.19.0",
    "cryptography>=42.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.1",
    "python-multipart>=0.0.6",
    "openai>=1.12.0",
    "anthropic>=0.40.0",
    "pyjwt>=2.8.0",
    "bcrypt>=4.1.0",
    "pymupdf>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
# Needed by pytest-asyncio 0.24+ to stop emitting the default-fixture-loop-scope
# deprecation warning. We run everything under a session-scoped loop so that
# session-scoped async fixtures (setup_db, test_client) can await work during
# startup/teardown without spawning a new loop per test.
asyncio_default_fixture_loop_scope = "session"
